requirements.txt
fastapi==0.135.1
uvicorn[standard]>=0.41.0
pydantic==2.12.5
google-genai==1.65.0
python-dotenv==1.2.2
requests==2.32.5
moviepy==2.2.1
google-cloud-storage==3.9.0
